requirements.txt
instagrapi
pillow
python-dotenv
wheel